+++ src/OperationalInsights/OperationalInsights/ChangeLog.md
@@ -19,6 +19,9 @@
 -->
 ## Upcoming Release
 * Fixed miscellaneous typos across module
+* Updated documentation for `New-AzOperationalInsightsLinuxSyslogDataSource`
+    - Added example
+    - Updated description for `-Name` parameter
 * Added an example for New-AzOperationalInsightsWindowsEventDataSource
 * Changed the description of the -Name parameter for New-AzOperationalInsightsWindowsEventDataSource